+++ packages/orq-rc/RELEASES.md
@@ -7759,3 +7759,13 @@ Based on:
 - [typescript v4.7.0-rc.19] packages/orq-rc
 ### Releases
 - [NPM v4.7.0-rc.19] https://www.npmjs.com/package/@orq-ai/node/v/4.7.0-rc.19 - packages/orq-rc
+
+## 2026-03-25 09:19:45
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.759.2 (2.869.23) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.7.0-rc.20] packages/orq-rc
+### Releases
+- [NPM v4.7.0-rc.20] https://www.npmjs.com/package/@orq-ai/node/v/4.7.0-rc.20 - packages/orq-rc